backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
groq==0.4.2
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.1.0
numpy==1.26.2
scikit-learn==1.3.2
rapidapi-connect==0.1.0